requirements.txt
python==3.9.21
altair==4.2.2
numpy==2.0.2
openai==1.73.0
pandas==2.2.3
plotly==5.24.1
python_dateutil==2.9.0.post0
Requests==2.32.3
streamlit==1.44.1
